refactoring for more consistent package name
